# libct/cg/sd: optimize and test findDeviceGroup