~~_This includes/depends on #3356, draft until merged._~~

1. Using strings.TrimPrefix instead of fmt.Sscanf.

2. Add a test case and a benchmark.

The benchmark shows some improvement, compared to the old
implementation:
```
    name               old time/op    new time/op    delta
    FindDeviceGroup-4    39.7µs ± 2%    26.8µs ± 2%  -32.63%  (p=0.008 n=5+5)
    
    name               old alloc/op   new alloc/op   delta
    FindDeviceGroup-4    6.08kB ± 0%    4.23kB ± 0%  -30.39%  (p=0.008 n=5+5)
    
    name               old allocs/op  new allocs/op  delta
    FindDeviceGroup-4       117 ± 0%         6 ± 0%  -94.87%  (p=0.008 n=5+5)

```