@kolyshkin 2022-03-17
Rebased, dropped the commit which contained optimization which did not resulted in any noticeable improvement.

@cyphar @thaJeztah PTAL