@kolyshkin 2022-03-17
@thaJeztah sorry to bother you, I have just updated the commit to remove most of bytes-related ugliness. Apparently, in modern Go versions, using bytes instead of strings does not result in any performance gain (it used to be the case, but appears to be fixed now).

The performance gain (and drastically reduced number of allocations) comes from ditching fmt.Sscanf which is apparently not implemented very well.

The new patch is much smaller and focused, PTAL.